+++ slsSupportLib/src/UdpRxSocket.cpp
@@ -15,7 +15,7 @@
 
 namespace sls {
 
-UdpRxSocket::UdpRxSocket(int port, ssize_t packet_size, const char *hostname,
+UdpRxSocket::UdpRxSocket(uint16_t port, ssize_t packet_size, const char *hostname,
                          int kernel_buffer_size)
     : packet_size_(packet_size) {
     struct addrinfo hints {};
